+++ src/seppl/__init__.py
@@ -6,3 +6,4 @@ from ._help import generate_plugin_usage, generate_help, HELP_FORMATS, HELP_FORM
 from ._types import get_class, get_class_name, classes_to_str
 from ._metadata import MetaDataHandler, add_metadata, get_metadata
 from ._split import gcd, Splitter
+from ._session import Session, SessionHandler